Update patch set 4
Patch Set 4:

We definitly have an issue with the 2 views fighting for the foreground. Both the background and the layout are being updated but because postInvalidate() is not instant we have a race to the foreground. This really should have been an issue sooner, but for some reason my older code was not updating the background consistently(which the code does properly now). We may want to consider only drawing the background on initiation, and adding a broadcast receiver for battery status changes to trigger a redraw of the background.

Patch-set: 4