Update patch set 4
Patch Set 4:

Just the flag would be probably "close enough" but it would still be possible to crash if the parallel execution of code was:

thread #1: in listener read flag == false
thread #2: in onDetached____ sets the flag and unregisters the listener service
thread #1: reads the notifications

While unlikely, I still don't like the possibility of seeing crashes in the logs.  At the same time, I added the flag in addition to the try and catch because I had the idea of losing exceptions.

Patch-set: 4