Update patch set 1
Patch Set 1:

A video says more than a thousand words...

https://drive.google.com/file/d/0B4Gcy6G7cgo8a0JIeHVQNFFpS2c/view?usp=sharing

Also, an issue I noticed, not related to this patch though. Sometimes the blue rectangle is seen when opening the cover, right before the lockscreen appears.

https://drive.google.com/file/d/0B4Gcy6G7cgo8blBtZ1BoUG02bXM/view?usp=sharing

I'm also wondering if the clock should be a little bit bigger and moved down a little, unless something else is supposed to occupy the space there.

Patch-set: 1