Update patch set 7
Patch Set 7: Code-Review-1

You are using the "timeout==0" as an effective OFF switch for the app by letting it block incoming intents. That blocks the app from actually handling the LID_CLOSED event and turning the screen off. With this change and timeout set to zero the screen is not turning off at all until the normal background timeout occurs.

Patch-set: 7
Label: Code-Review=-1