Update patch set 1
Patch Set 1:

Default is the same as it has been until now. The outline option is for the ZUK Z2 where the official case has a clear plastic ring around the circular cutout, and having a matching ring of colour on the LCD makes it look like the plastic ring is glowing.

Patch-set: 1