[WIP]FlipFlap: Constrict Layout to the same dimensions as view window
We have been assuming all view windows started at the top of the screen.
This should not be expected, there are devices with oddly positioned view
windows. Restrict the layout dimensions to the same size/position as the
given view window.

TODO: Circle window still needs work for this, more to come

Change-Id: Ib5ba8fbc777b03ec5bb563bb9c0bf1a67ea4a8f4