Update patch set 1
Patch Set 1:

The problems were observed on a Galaxy S5 with S-view cover.
The patches were not tested yet on a real device, but we tested on a demo application :D
This demo is derived from this application, with the protected api removed so it can be installed on emulator, and it have a new activity where I can add a rectangularView or a circleView. The patch can be a little risky, but the other ones do not touch the layout

Patch-set: 1
Reviewer: Florin9doi <20248@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>